Added music to the title scene. Need to make intro play seemlessly into loop
