+++ Assets/Scripts/Core/MetaGameController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 86f4a4ae3a36547dd915a91ceebed970
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
