+++ Assets/Scripts/Core/TitleSceneController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 38200764171f17e418142a22b20bcc19
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
